Update chart from Listener pid
Each metric projection (with the exception of LastValue) will need
access to some state, which we can store in the Listener.

The Process.sleep() calls in the tests are unfortunate but they
should suffice until we have a better way to ensure that the chart
has received all of its events before asserting on the connect()
data.